Add new entry methods to GlobalCaches
Add `compute_size_for_memory_monitor(double)` to both the GlobalCache
and the MutableGlobalCache. This new entry method is used for
calculating the size of the caches for the MemoryMonitor. They can
only be run if the MemoryMonitor component is in the component list
in the metavaraibles

Add `set_global_cache_proxy(CProxy_GlobalCache<Metavariables>)` to
MutableGlobalCache and call this in the GlobalCache constructor.
The previous entry method calls a simple action.
On the GlobalCache this isn't an issue because we can get the local
GlobalCache by doing this->thisProxy. But the MutableGlobalCache is
unaware of the GlobalCache. We cannot pass a GlobalCache& to the previous
entry method because charm requries reference arguments to entry methods
be const. However, to call a simple action, we can't have a const&, we
need a non-const&. Thus, the solution was to have the MutableGlobalCache
hold a proxy to the GlobalCache internally and since the
MutableGlobalCache is constructed before the GlobalCache, this had to be
done after (in the GlobalCache constructor) with an entry method that
will broadcast to all branches of the MutableGlobalCache.

fixup global cache entry method