Add self start Action to check time stepper history
Put this check in a new phase CheckTimeStepperHistory

This Action is intended to be run in the CheckTimeStepperHistory
phase. It will check that we have the proper number of entries
in our history, and that our TimeStepId is not in self-start
mode. If either of those are true, than an error will occur and
the evolution will not proceed