WIP. Use Charm messages
I think I may need to edit the GhostData class yet again. Currently
in CTD, we don't have anywhere to store the subcell data. I had
intended to use GhostDataForReconstruction for this, but it's the
wrong data type. The subcell data to send is in DirectionMap while
the type upon receive is map<pair<Direction, ElementId>, GhostData>.

Ideally, I think we should create a custom type (similar to GhostData)
for everything. Internally it would store a DirectionMap for the
local data, and then a map<pair<Direction, ElementId>, GhostData> for
the received data. That way all subcell data is managed in a single
place.